+++ backend/database/migrations/2025_10_15_170609_create_manipulate_cost_per_unit_materials_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('manipulate_cost_per_unit_materials', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('unitCost');
+            $table->decimal('unitCost');
             $table->timestamps();
         });
     }
